# internal/auth/delivery/http/models.gen.go
// Package v2 provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.12.4 DO NOT EDIT.
package v2

// ErrorHTTP defines model for ErrorHTTP.
type ErrorHTTP struct {
	// Code Error code
	Code int32 `json:"code"`

	// Message Error message
	Message string `json:"message"`
}

// JsonWebToken JWT иокен
type JsonWebToken = string

// SignUp defines model for SignUp.
type SignUp struct {
	// Login Логин.
	Login string `json:"Login"`

	// Password Пароль.
	Password string `json:"Password"`
}

// LoginJSONRequestBody defines body for Login for application/json ContentType.
type LoginJSONRequestBody = SignUp