# internal/user/delivery/http/models.gen.go
// Package v2 provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.12.4 DO NOT EDIT.
package v2

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// CreateUser defines model for CreateUser.
type CreateUser struct {
	// Avatar Аватар пользователя.
	Avatar string `json:"Avatar"`

	// Email Email пользователя.
	Email string `json:"Email"`

	// Login Логин пользователя.
	Login string `json:"Login"`

	// Name Имя пользователя.
	Name string `json:"Name"`

	// Password Пароль пользователя.
	Password string `json:"Password"`

	// PasswordCheck Проверка пароля пользователя.
	PasswordCheck string `json:"PasswordCheck"`

	// PhoneNumber Номер телефона пользователя.
	PhoneNumber string `json:"PhoneNumber"`
}

// ErrorHTTP defines model for ErrorHTTP.
type ErrorHTTP struct {
	// Code Error code
	Code int32 `json:"code"`

	// Message Error message
	Message string `json:"message"`
}

// GetUser defines model for GetUser.
type GetUser struct {
	// Avatar Аватар пользователя.
	Avatar string `json:"Avatar"`

	// ID Логин пользователя.
	ID string `json:"ID"`

	// Name Имя пользователя.
	Name string `json:"Name"`

	// PhoneNumber Номер телефона пользователя.
	PhoneNumber string `json:"PhoneNumber"`
}

// JsonWebToken JWT иокен
type JsonWebToken = string

// CreateUserJSONRequestBody defines body for CreateUser for application/json ContentType.
type CreateUserJSONRequestBody = CreateUser

// UpdateUserJSONRequestBody defines body for UpdateUser for application/json ContentType.
type UpdateUserJSONRequestBody = CreateUser